整理：impl Transcriber for ResidentWhisper 迁回 whisper_server.rs
transcriber.rs 与 whisper_server.rs 互 use 形成双向耦合：
- whisper_server.rs 反向 use transcriber::LocalWhisper 作回退
- transcriber.rs:469 写 trait impl for whisper_server::ResidentWhisper

impl 块迁回 whisper_server.rs（trait 类型通过 use crate::transcriber
显式引入），transcriber.rs 删除该 impl 与对应的反身 use。
模块边界变清晰：transcriber.rs 定义 Transcriber trait 与 LocalWhisper，
whisper_server.rs 定义 ResidentWhisper 并自己实现 trait。
